==> NoteBook/02_Funciones, For loops.Rmd <==
---
title: "Funciones, For loops, If else"
output: html_notebook
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Ctrl+Shift+Enter*. 

```{r}
plot(cars)
```

Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Ctrl+Alt+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Ctrl+Shift+K* to preview the HTML file).

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.


_INTRODUCCIÓN_

Para la clase de hoy vamos a ver comandos en RStudio que puede ayudar a automatizar nuestro codigo condensando pasos repetitivos en un solo comando (pero varias lineas de codigo) 

##Funciones

Set de argumentos que se organizan juntos para realizar una tarea especifica. 

En realidad todos los comandos que hemos usado son la funciones que ya vienen incluidos en R pero RStudio tambien nos da la posibilidad de crear nuestras pripias funciones. 

###Componentes de una funcion

*__Nombre de la funcion:__Nosotros se la damos
*__Argumentos:__Los elementos necesarios para llevar a cabo la funcion. _input_
*__Cuerpo de la función:__Define lo que va a hacer la funcion
*__Respuesta:__El valor o resultado de la tarea asignada a una funcion _output_


```{r}
nombre_funcion <- function(argumentos) {#es lo que necesitas para calcular la funcion 
respuesta <- haz_algo (argumentos) #calculos o modificaciones que haras con los argumentos
return(respuesta) #lo que se va a "imprimir"} #todo loq ue esta entre corchetes es la funcion

```

Piensen en las funciones como una accion por lo general el nombre empieza con un verbo: 


```{r}
calc_vol_foliar <-function(longit, ancho, altura) {
  area <- longit * ancho
  volumen <- area * altura
  return(volumen)
  }

#Ahora usemos la funcion

calc_vol_foliar(longit = 1,
                ancho = 2.5,
                altura = 2)

#hay que tener muy encuanta el orden de la longitud, ancho y altura
#tambien hay que tener en cuanta que el resultado me va a aparecer en la consola pero no me va a aparecer en el environment ya que no colocamos un objeto antes, si queremos guardar el dato tenddríamos que hacer lo siguiente: 

arbusto1 <- calc_vol_foliar(longit = 1,
                            ancho = 2.5,
                            altura = 2)


#ya con el echo de colocar antes un objeto se te va a guardar el dato y te aparecera no solo en la consola sino tambien en el apartado del environment
#tambien es importante que llevemos a cabo el orden de los corchetes



#Ahora usemos datos desde un dataframe

###Crearemos un set de datos artificiales###


df <- data.frame(arbustos= c("arbusto1", "arbusto2"),
                 longitud= c(20, 17), 
                 ancho= c(11, 9),
                 altura= c(30, 28))

#esto fue un ejemplo creado por mi 

#ejemplo grupal 

calc_vol_foliar(df$longitud[2], #en estos brackets no hay como porque es una columna osea solo una dimension
     df$ancho[1],
     df$altura[2])

#sin la funcion

df$longitud[1]*df$ancho[1]*df$altura[1]

#valores default

calc_vol_foliar <- function(longitud, ancho, altura = 1) {
  area <- longitud * ancho
  volumen <- area * altura
  return(volumen)
  }

calc_vol_foliar(longitud = 3,
                 ancho = 2.5)

#valores default igual se pueden modificar 

calc_vol_foliar(longitud = 4,
                ancho = 2.5,
                altura = 3)


```



_EJERCICIO_

van a crear una funcion que calcule el area foliar de una especie de arboles unica en el mundo que tiene hojas triangulares; dato el area de un triangulo es base por altura entre 2 (Cr = A ^1.73). 

```{r}
calcular_area_foliar <- function(base, altura){
  area <-(base * altura) / 2
  return(area)
}

calcular_area_foliar(base = 3, altura = 5)


#FUNCION PARA ESTIMAR CRECIMIENTO DE LAS HOJAS 

est_crecim <- function(area) {
  crecim <- area ^1.73
  return(crecim)
}

##FUNCIONES COMBINADAS O NESTED FUNCTIONS

crecim_hoja <- est_crecim(calcular_area_foliar(3,5))
```

_FOR LOOPS_

Los For Loops se usan cuando hay un pedazo de codigo que necesitan repetir varias veces

#Structura basica
```{r}
for (numero_de_cosa in lista_de_cosas) {
  accion(cosa)
}

```

###Ejemplo simple para for loops
```{r}
longits <- c(1.6, 3, 5, 19) #vector

for (a in longits) { # el nombre a es arbitrario muy bien podria usar solo una letra u otro nombre
  print(2.65*a ^ 0.9) #pero el nombre debe coincidir con la accion que se esta comandando
} # se pone print para obtener el resultado y poder verlo pero no esta guardado

## un for loop con multiples pasos
for (longit in longits) {
  masa <- 2.65 * longit ^ 0.9
  masa_lb <- masa * 2.2
  print(masa_lb)
}
```
###Ejercicio

```{r}
#Vamos a modificar este for loop para que imprima multiplos de 3 

for (i in 1:5) {
  print(i * 3)
}


##Ahora hacer un for loop para cambiar estos valores de centimetros a pulgadas cm/2.54
```


```{r}
centim <- c(135,45,23,167,30,54,29)

for (cm in centim) {
  pulgadas <- cm / 2.54
  print(pulgadas)
}

```

###Como se aplica un for loop a un dataframe

Para  poder seleccionar una de las columnas o filas se elige un index o posicion para que se vaya repitiendo

```{r}
Medidas_clase <- 
  read.csv("~/Rstudio/CursoInnovak/Materiales/Class-data.csv")

#Modificar la altura de cm a pies

for (i in 1 : nrow(Medidas_clase)) {
  altura_ft <- Medidas_clase$Height.cm[i]/30.48
  print(altura_ft)
}
```
si bien asi pudimos transformar toda la altura a pies, en realidad la tabla sigue sin haberse cambiado y esos resultados se imprimieron 

COMO PODEMOS GUARDARLO 

```{r}
#primer loop
alt_pies <- vector(mode = "numeric",
                   length = nrow(Medidas_clase))

#loop que guarda los datos
for (i in 1: nrow(Medidas_clase)) {
  altura_ft <- Medidas_clase$Height.cm[i]/30.48
  alt_pies[i] <- altura_ft #cambiamos print para que lo guarde en el vector
}

alt_pies

```
###Ejercicio

Ahora repetir el proceso de guardar los datos de un for loop dentro de un vector modificando el for loop que usamos para modificar centimetros a pulgadas

```{r}
centim <- c(135,45,23,167,30,54,29)

alt_pulgadas <- vector(mode = "numeric",
                   length = length(centim))

for (e in 1: length(centim)) {
  pulgadas <- e / 2.54
  alt_pulgadas[e] <- pulgadas
}

```

###Integrar datos de un for loop en un data frame

```{r}
#Unir el vector con el dataframe "forma fácil"

Medidas_clase <- cbind(Medidas_clase,alt_pies) #la C es de columna, lo que significa que quiero que en el documento de medidas_clase al final me ponga una columna que se llame alt_pies (cbind significa= c de columna y bind de unir)

#Directamente en la tabla "forma dificil"

for (i in 1: nrow(Medidas_clase)) {
  altura_ft <- Medidas_clase$Height.cm[i]/30.48
  Medidas_clase$alt_pies[i] <- altura_ft #cambiamos para que lo guarde directamente en el dataframe ZONA DE GUARDADO
}


##Hacer calculos con las columnas 
Acidos <- read.csv("~/RStudio/CursoInnovak/Materiales/Acidos_ex.csv")

sapply(Acidos, mode) #Esto es para checar el tipo de caracter de cada columna en una data frame

for (i in 2: ncol(Acidos)) # yo nombro mi objeto como i, y quiero que en la columna nuemro de mi documento acidos me saque el promedio {
  Promedios <- mean(Acidos[,i], na.rm = TRUE) 
  print(Promedios)
}

mean(Acidos[,2], na.ra = TRUE) #esta formula me dice que en mi columna nuemro 2 de mi archivo llamado acidos me quite o no tome en cuenta mis casillas que no hay datos. EJEMPLO DEL CODIGO SOLITO

#Sin embargo este no se guardo por lo que este seria el proceso para guardarlo 

ac_prom <- c() 

for (i in 2:ncol(Acidos)) {
  Promedios <- mean(Acidos[,i], na.rm = TRUE)
  ac_prom[i] <- c(Promedios)
}

Acidos <- rbind(Acidos,ac_prom) #row bind eso quiere decir que estara hasta abajo
```

####Ejercicios

```{r}
# 1. Agregar al archivo de Medidas_clase otra columna en el que la circunferencia de la cabeza este en pulgadas (este es con nrow) y guardarlo en el archivo, recordar que cm/2.54

for (i in 1: nrow(Medidas_clase)) {
  circun_in <- Medidas_clase$CircumferenceHead.com[i]/2.54
  Medidas_clase$circun_in[i] <- circun_in
}



# 2. En el archivo de acidos ya agregamos una fila con los promedios ahora traten de agregar una con la desviacion estandar (en este afuerzas como el ejercicio de arriba)

Desv_ac <- c() 

for (i in 2:ncol(Acidos)) {
  Desv_std <- sd(Acidos[,i], na.rm = TRUE)
  Desv_ac[i] <- c(Desv_std)
}

Acidos <- rbind(Acidos,Desv_ac)
```

_COMANDOS If Else_

Los comandos de If Else en RStudio son una parte fundamental de la programacion ya que la computadora es capaz de "tomar decisiones" dandole _condiciones_

###El mas basico: If Statment

```{r}
#Estructura basica IF Statement

if (Condicion_que_debe_ser_verdad) {
  Haz_algo
}

##Ejemplo con condiciones numericas 

x <- 7

if (x > 5){ #estos objetos si aparecen en el ambiente global
  x <- x ^ 2 #si (If) es FALSO no hace nada
}

#Ejemplo biologico 

veg_tipo <- "arbol"
volumen <- 16

if (veg_tipo == "arbol") {#cuando pones solo un = es una operador matematico, mientras que == significa es igual a 
  masa <- 2.6 * volumen ^ 0.9 #este objeto si se guarda
}

masa

```
#EJERCICIO

Como maestros de secundaria ustedes estan calificando unos examenes. Si el alumno tuvo mas de 70 aciertos en el examen pasa. Creen un _if statement_ que me arroje APROBADO dependiendo de la clasificacion del alumno

```{r}

aciertos <- 71

if (aciertos > 70) {
  print("APROBADO")
}

```
###Complicamos mas la cosa: If Statement

Con el _If statement_ solo le decimos a la computadora si los datos que le dimos son CIERTOS si son FALSOS este comando simplemente no arroja nada por lo que le tenemos que agregar el _Else_

```{r}
x <- 4

if (x > 5) { 
  x <- x ^ 2 #asignar pero no se muestra
} else { #ahora si (If) es falto tambien le decimo sque hacer
  x <- x ^ 3
  }

#Y se puede complicar mas la cosa

veg_tipo <- "pasto"
volumen <- 16

if (veg_tipo == "arbol") {
  masa <- 2.6 * volumen ^ 0.9 #este objeto si se guarda
} else if (veg_tipo == "pasto") { #cadenita de condiciones 
  mass <- 0.6 * volumen ^ 1.2
} else {#agregar la ultima condicion
  mass <- NA
  }

```

####Ejercicio 

Como maestros de secundaria ustedes siguen calificando los examenes pero ahora a los alumnos que sacaron mas de 90 aciertos les va a poner EXCELENTE; si tuvieron entre 90 y 70 aciertos dira APROBADO y menos de 70 REPROBADO. Modifiquen su codigo de _If statement_ agregando _else_ para que me arroje las tres categorias dependiendo de la calificacion del alumno

```{r}
aciertos <- 95

if (aciertos > 90) {
  print("EXCELENTE") 
} else if (aciertos >= 70) {
  print("APROBADO")
} else {
  print("REPROBADO")
}

```

###La version simplificada: ifelse

Esta version se define por estar "vectorizada" por lo que en vez de tener las condiciones en renglones, todo esta condensado en una linea

Es util cuando cuando las comparaciones son mas sencillas e interesantemente se puede usar directamente con vectores en vez de valores simples 

```{r}
#Estructura simplificada

ifelse(condicion, respuesta_TRUE, respuesta_FALSE)

#Ejemplo basico 

edad_planta <- 12

ifelse(edad_planta < 2, "Plantula", "Adulto")

#Ejemplo anidado

ifelse(edad_planta < 2, "Plantula", ifelse(edad_planta < 7, "Juvenil", "Adulto"))


#o puede ser por separado en lugar de en linea de la siguiente manera 
ifelse(edad_planta < 2, #condicion
       "Plantula", #verdadero)
       "Adulto") #falso
ifelse(edad_planta < 2, #condicion
       "Plantula", #verdadera
       ifelse(edad_planta < 7, "Juvenil", "Adulto")) #falsa

```
###Ejercicio

El codigo previo que hiciero usando If Else ahora conviertanlo al comando _ifelse_ 

*pista debe de se4r anidado

```{r}
aciertos <- 73

ifelse( aciertos > 90, "Excelente", ifelse(aciertos > 70, "Aprobado", "Reprobado"))

```
###Aplicando ifelse a vectores

```{r}
# vector de numeros

a <- c(2,5,6,7)

#checar si es par o impar 

ifelse(a %% 2 == 0, "Par", "Impar")

#condicion %% 2 == 0, cuando divides entre 2 el residuo es cero

```
