==> NoteBook/04_Estadistica descriptiva y normalizacion de datos.Rmd <==
---
title: "Estadistica descriptiva y normalizacion de datos"
output: html_notebook
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Ctrl+Shift+Enter*. 

```{r}
plot(cars)
```

Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Ctrl+Alt+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Ctrl+Shift+K* to preview the HTML file).

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.

# Introduccion 

Ahora que ya sabemos como procesar y modificar nuestras tablas, es momento de aplicar todo lo aprendido a una de las bases del metodo cientifico: __el analisis estadistico__ 

Si bien hay muchas maneras de analizar datos y aun mas metodos para ello. A lo largo de este modulo vamos a aprender un flujo basico para determinar que tipo de analisis es el mas idoneo para cada tipo de datos.

```{r}
# Libraries

library(dplyr) # manipulacion de datos, funciones, modificar columans, filtrar, reordenar, calculos, etc 
library(tidyr) # modificando las funciones de datos
library(readr) # leer y escribir csvs
library(udunits2) # conversion de unidades 


# Data

Plantas <- read.csv("~/RStudio/CursoInnovak/Materiales/Tabla_plantas.csv")

Stream <- read.csv("~/Rstudio/CursoInnovak/Materiales/stream.csv")

Acidos <- read.csv("~/RStudio/CursoInnovak/Materiales/Acidos_ex.csv")

Fosf <- read.csv("~/RStudio/CursoInnovak/Materiales/phosphorus.csv")

Medidas_clase <- read.csv("~/RStudio/CursoInnovak/Materiales/Class-data.csv")


```

# Estadistica descriptiva

En la clase anterior ya vimos, los valores mas importantes dentro de la estadistica descriptiva. El objetivo de la estadística descriptiva es describir los datos observados de forma sintética y significativa para poder analizarlos mejor. La estadística descriptiva comprende tres categorías principales: Te describe los datos observados de manera simple pero no te dira cual es mejor que otro.

1. Distribución de frecuencias
    + Count: n(), n_distinct() en dplyr o count()
    + Range: min(), max(), quantile()
2. Medidas de tendencia central 
    + Center: mean() _que es el promedio_, median() _que es la mediana se usa cuando hay una dispersion muy alta_ 
3. Medidas de variabilidad.
    + Spread que es dispersion: sd() _que es desviacion estandar si usamos media_, IQR() _que es el rango intercuartilico_, mad() _si cualculas mediana usar el mad y no la desviacion estandar_

```{r}
## 1. Distribucion 

# Count
count(Medidas_clase,sex, Handedness)

# Range
min(Medidas_clase$Height.cm)
max(Medidas_clase$Height.cm)
quantile(Medidas_clase$Height.cm)

## 2. Tendencia central
mean(Medidas_clase$Height.cm) # si tienes celdas vacias no olvides el na.rm = TRUE
median(Medidas_clase$Height.cm)

## 3. Variabilidad 
sd(Medidas_clase$Height.cm)
IQR(Medidas_clase$Height.cm)
mad(Medidas_clase$Height.cm)

```
# Ejercicio 

Usando el data fame de plantas cada quien va a elegir una variable y comparar media, mediana, desviacion estandar y desviacion de la mediana

```{r}
mean(Plantas$root_tip_count, na.rm = TRUE)
median(Plantas$root_tip_count, na.rm = TRUE)

sd(Plantas$root_tip_count, na.rm = TRUE)
mad(Plantas$root_tip_count, na.rm = TRUE)

```

# Normalizacion de datos 

La distribucion nomral de datos se caracteriza por gran cantidad de datos alredador de la media y pocos datos en las colas de forma _simetrica_. 

Es decir si conocemos la media y la desviafcion estandar son conocidos la probabilidad de que cualqueir obsrvacion futura caiga dentro de un rango especifico ya es conocida. 

Todos los anaisis parametricos tiene como primera suposicion que los datos que se van a usar en el analisis siguien una distribucion normal e ahi la importancia de normalizar los datos. 

si bien existen tests para datos no parametricos estos son mas estrictos y no siempre van a arrojar diferencias a pesar de que si las hayan. 

## Como saber si nuestros datos son normales? 

```{r}
# Visualmente 

hist(Stream$Temp.degC)
qqnorm(Stream$Temp.degC)

# Shapiro-wilk normality test _tiene que ser mayor a 0.05 para ser normal_ te mide que tus datos no sean normales por lo tanto si son mayores quiere decir que es verdad que si son normales.

shapiro.test(Stream$Temp.degC)
```
El test de shapiro es muy estricto por lo que datos que tengan un p-value cerca de 0.05 y un histograma y qqplot mas o menos lineal pueden ser considerados para estadisticas parametricas. Por ello cuando trabajamos con datos hay que empezar revisando el test shapiro y de ahi continuar el analisis para deterimar si el set de datos es normal. 

## Ejercicio 

En el data frame _plantas_ cada quien elija una variale y determinen si siguien o no una distribucion normal. JUSTIFICAR 

```{r}
par(mfrow=c(1,2)) #este codigo es para mostrar los graficos juntos 
hist(Plantas$root_tip_count)
qqnorm(Plantas$root_tip_count)

shapiro.test(Plantas$root_tip_count)

```

### Como checar normalidad en muchos datos

```{r}
row.names(Acidos) <- Acidos[,1]
Acidos <- Acidos[,-1]
Acidos <- as.data.frame(t(Acidos))

for (i in 1:ncol(Acidos)) {
  shapiro <- shapiro.test(Acidos[,i])
  normal <- ifelse(shapiro[("p.value")]>0.05, "YES", "NO")
  print(c(i,normal))
}

for (i in 1:ncol(Acidos)) {
  hist(Acidos[,i],
       main = i)
}

# como harias el qqplot? 

for (i in 1:ncol(Acidos)) {
  qqnorm(Acidos[,i],
       main = i)
}
```

### Ejercicio

Modifiquen el codigo anterior para checar normalidad en todas las variables del dataframe _plantas_

```{r}
for (i in 6:ncol(Plantas)) {
  shapiro <- shapiro.test(Plantas[,i])
  normal <- ifelse(shapiro[("p.value")]>0.05, "YES", "NO")
  print(c(i,normal))
}


for (i in 6:ncol(Plantas)) {
  hist(Plantas[,i],
       main = i)
}

# como harias el qqplot? 

for (i in 6:ncol(Plantas)) {
  qqnorm(Plantas[,i],
       main = i)
}
```

## Como transformar datos para que sean normales 
La transformacion aplicada dependera del tipo de datos y el tipo de distribucion o sesgo que tenga: 

* Sesgo a la izquierda = esta la colita a la izquierda
* Sesgo al a derecha = la colita esta a la derecha 
* Sesgo uniforme = todo esta al mismo nivel 
* Sesgo simetrical = como una montaña
* Sesgo bimodal = tienen dos picos es la mas dificil de modificar

Aunque existen ecuaciones mas complejas para sets de datos especificos, no es recomendable transformar los datos hasta que se vuelven numeros artificiales que representan verdaderas diferencias biologicas. Por ello si despues de dos metodos de transformacion son usados y los datos siguen sin ser normales lo mejor es usar un metodo no parametrico para analizarlos. 

```{r}
# Checar normalidad

shapiro.test(Fosf$TDP)
hist(Fosf$TDP)
qqnorm(Fosf$TDP)
#estos datos no son normales


# Transformar usando log

shapiro.test(log(Fosf$TDP)) # log() significa logaritmo natural
hist(log(Fosf$TDP))
qqnorm(log(Fosf$TDP))

### Como transformar de manera grupal

Acid_log <- Acidos

for (i in 1:ncol(Acidos)) {
  Acid_log[,i] <- abs(log(Acidos[,i])) #abs es valor absoluto
}
```

### Ejercicio final

De las variables que concluimos no eran normales en el dataframe _platas_ elegir el mejor metodo de transformacion para normalizarlos y asi poder analizarlos. Y enseñar shapiro, histograma, qqplot antes y despues de normalizar. 

```{r}
shapiro.test(Plantas$Roots)
hist(Plantas$Roots)
qqnorm(Plantas$Roots)

# Transformar usando log, se utilizo el log porque es un sesgo al a derecha 

shapiro.test(log(Plantas$Roots)) # log() significa logaritmo natural
hist(log(Plantas$Roots))
qqnorm(log(Plantas$Roots))

```
