==> NoteBook/06_ANOVA y Tukey HSD.Rmd <==
---
title: "ANOVA y Tukey HSD"
output: html_notebook
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Ctrl+Shift+Enter*. 

```{r}
plot(cars)
```

Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Ctrl+Alt+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Ctrl+Shift+K* to preview the HTML file).

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.



# INTRODUCCION

A un nivel mayor de complejidad que las pruebas de T el ANOVA (Analysis of variance) es una formula estadistica que se utiliza para comparar varianzas entre las medias de diferencia (_mas de dos_) grupos.

Existe el ANOVA de una via en la que solo se tiene una variable independiente ( con varios grupos) y se quiere ver la diferencia entre dichos grupos sobre una variable dependiente. 

El ANOVA de 2 vias añade otra variable independiente por lo que mide el efecto de dos variables independientes sobre una variable dependiente. Tambien existe la ANOVA de tres vias pero la complejidad en comparaciones y tratamientos crece bastate. 

```{r}
# Library

library(dplyr)
library(tidyverse)
library(car)
library(agricolae)


# DATA

Rodillas <- read.csv("~/RStudio/CursoInnovak/Materiales/Rodillas.csv")

Estanques <- read.csv("~/RStudio/CursoInnovak/Materiales/BRW.ponds.csv")

Comp_suelos <- read.csv("~/RStudio/CursoInnovak/Materiales/Soil_Comparison.csv")

```


# ANOVA de una via

Es decir solo UN tratamiento independiente 

# ANOVA con R base

R base puede calcular ANOVA de dos maneras diferentes, una de ellas usando el codigo _anova()_. Sin embargo, esta funcion base calcula automaticamente una ANOVA tipo I la cual da por sentado que los datos son balanceados por lo que cuando no sea asi se debe de usar la funcion del paquete _car_

```{r}

# variable dep ~ variable indep, datos
summary(aov(shift~treatment, data = Rodillas)) #aov es anova

anova(lm(shift~treatment, data = Rodillas)) #lm quiere decir modelo lineal, es el recomendable para usar. puedes cambiar el lm por aov ya que es lo mismo pero pueden llegar a tener difernecias por lo que mejor usar el lm.
```
La diferencia entre lm y aov es el tipo de modelo y mas que nada la presentacion de los datos. Si quieren sabedr mas al respecto pueden leer aqui: 

pegar liga 

sin embargo, para fines utiles siempre usar el aov() ya que por definicion significa analisis de varianza 

# ANOVA con el paquete CAR

Este paquete esta enfocado a varios tipos de analisis estadisticos y su funcion de ANOVA tiene un argumento para lidiar con datos desbalanceados (Tipo II y III). Ademas que este paquete tambien cuenta con la funcion _levene_ para checar igualdad de varianzas. 
*si no tienen correlacion o son desbalanceados usar el paquete tipo II
pegar liga

```{r}
# Revisar por homogeneidad de varianzas 

leveneTest(shift~treatment, data = Rodillas) # mayor a 0.05, no utilizar levene.test porque no sirve 

plot(rod_anova,1) # Interpretacion: los residuals tienen que estar distribuidos de manera aleatoria y tienen varianza constante. Este casi no lo vamos a ocupar. Solo se ocuparia si el levene no pasa.

## ANOVA

rod_anova <- aov(shift~treatment, data = Rodillas) # se puede usar lm o aov pero mejor aov

Anova(rod_anova) # fijarse que va con mayuscula 

#automaticamente da el tipo II

## ANOVA tipo III
Anova(rod_anova, type=3) # esge no es el correcto de correr para este set de datos 

```
La diferencia principal entre el tipo II y tipo III es la interaccion de los grupos de las variables independientes. Por ejemplo es recomendable usar tipo II si no estan seguros si sus variables interactucan entre si osea que esten desbalanceados 

_CONCLUSION_ si p<0.05 es diferente osea hay diferencias

# Tukey HSD

En el caso que una ANOVA nos muestre que existen diferencias significativas entre tratamientos, el paso siguiente es correr un test de Tukey HSD en el cual nos "desglosa" la ANOVA y nos dice cuales de los tratamientos son diferentes. 

Para esto solo hay un codigo el cual es _TukeyHSD()_

IMPORTANTE; el ANOVA te dice que si hay diferencias y el TUKEY te dice cuales son las diferencias. 

```{r}
TukeyHSD(rod_anova) # los que son diferentes son los que son menores a 0.05 en este caso los diferentes son los ojos-control y rodilla-ojos = p<0.05

TukeyHSD(aov(shift~treatment, data = Rodillas))
plot(TukeyHSD(rod_anova)) # si el intervalo de ocnfianza no se superpone con la diferencia media no son significativos

### en la grafica vemos que no se supoerponen en la media los ojos-contol y ojos rodillas por o cual son significativas eso quiere decir que son diferentes osea que hay diferencia significativa 


```

## Ejercicio 

Usando el dataframe de _Estanques_ vamos a comparar la calidad del agua en los estanques entre tres regiones diferentes de Alaska. Para ello cada uno de ustedes seleccionar una de las variables y llevar a cabo un test ANOVA. No olviden tomar en cuenta los pasos: 

* checar normalidad y de no estar estandarizado  transformar
* checar si el set de datos esta balanceado (para elegir el tipo de ANOVA)
*checar homogeneidad de varianza 
*Calcular ANOVA y si es significativa calcular Tukey HSD 
*Hacer un grafica del resultado Tukey

```{r}

# NORMALIDAD 
shapiro.test(Estanques$TP)

par(mfrow=c(1,2))
hist(Estanques$TP)
qqnorm(Estanques$TP)

# TRANSFORMADO
par(nfrow=c(1,2))
hist(log(Estanques$TP))
qqnorm(log(Estanques$TP))
shapiro.test(log(Estanques$TP))


# BALANCE; 
Estanques %>%
  group_by(Region) %>%
  summarise(n()) #asi es el codigo 

# HOMOGENEIDAD
leveneTest(log(TP)~Region, data = Estanques) # mayor a 0.05, una vez que transforme en shapiro se tiene que utilizar el log en homogeneidad 


## ANOVA
estanque_anova <- aov(log(TP)~Region, data = Estanques) 

Anova(estanque_anova) 

#CONCLUSION: p<0.05 por lo que hay diferencia significativa en los datos 


#TUKEY
TukeyHSD(estanque_anova) # los que son diferentes son los que son menores a 0.05 en este caso los diferentes son los BRW-BEO Y IBP-BRW = p<0.05

plot(TukeyHSD(estanque_anova)) # si el intervalo de ocnfianza no se superpone con la diferencia media no son significativos

### en la grafica vemos que no se supoerponen en la media los BRW-BEO y IBP-BRW por Lo cual son significativas eso quiere decir que son diferentes osea que hay diferencia significativa 


```
# IMPORTANTE SABER QUE 
Cuando hay datos con 0 los tomaria el programa como datos infinitos y al programa no le sirven esos datos por lo cual hay que ponerle un +1 a la columna que tiene esos datos y asi correra. 


# ANOVA DE 2 VIAS 

Aunque este tipo de analisis es sumamente util, muchas veces nuestros experimentos incluyen mas de una variable independiente que manipulamos. Para ello los datos se pueden comparar usando una ANOVA es de 2 o mas vias. 

```{r}
leveneTest(logNH3~Temperature*Soil, data = Comp_suelos)

#Balanceado?
Comp_suelos %>%
  group_by(Temperature,Soil) %>%
  summarise(n())

# ejecto aditivo

est_anova <- aov(logNH3 ~ Temperature+Soil, data = Comp_suelos)
Anova(est_anova)

#Interracion entre variables 
est_anova <- aov(logNH3 ~ Temperature*Soil, data = Comp_suelos)
Anova(est_anova)

#Tukey
est_Tukey <- TukeyHSD(est_anova)
plot(est_Tukey)

# El camino facil: Agricolae 

summary(est_anova)

rtout <- HSD.test(est_anova, trt = c("Temperature", "Soil"),
                  alpha = 0.05) #este es el codigo de tukey en la libreria agricolae, como es de dos filas en la C le pones cuales columnas y le pones el valor de significancia que es el 0.05

```

#### EJERCICIO

Ahora les toca a ustedes cada quien va a elegir 1 variable y comparar los efectos de temperatura y tipo de suelo en ellas. Tienen que seguir el mismo proceso y al final me van a explicar porque si o no hay diferencia significatica entre los grupos. 

```{r}
#Homogeneidad de datos
leveneTest(logTDP ~ Temperature*Soil, data = Comp_suelos) #mayor a 0.05

#Balance
Comp_suelos %>%
  group_by(Temperature,Soil) %>%
  summarise(n())

#Interaccion
TDP_anova <- aov(logTDP ~ Temperature*Soil, data = Comp_suelos)
Anova(TDP_anova)

#Tukey
TDP_Tukey <- TukeyHSD(TDP_anova)
plot(TDP_Tukey)

# El camino facil: Agricolae 

summary(TDP_anova)

r <- HSD.test(TDP_anova, trt = c("Temperature", "Soil"),
                  alpha = 0.05) #este es el codigo de tukey en la libreria agricolae, como es de dos filas en la C le pones cuales columnas y le pones el valor de significancia que es el 0.05

```
