==> NoteBook/07_Regresion y Correlacion.Rmd <==
---
title: "Regreciones y Correlaciones"
output: html_notebook
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Ctrl+Shift+Enter*. 

```{r}
plot(cars)
```

Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Ctrl+Alt+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Ctrl+Shift+K* to preview the HTML file).

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.


# Introduccion

Para el ultimo tema del curso vamos a ver como calcular regresiones y hacer correlaciones lineales en RStudio. Muchas veces confundidas entre si las regresiones y las correlaciones son las tecnicas mas comunes para determinar si existe una relacion entre dos variables cuantitativas.


```{r}
#Data
FlujosC <- read.csv("~/RStudio/CursoInnovak/Materiales/flujos.csv")

```

# Correlacion 

Por un lado, la  correlacion o mas propiamente dicho _el coeficiente de correlacion_ (hay varios metodos el mas comun es el de Pearson) _r_ cuantifica la fuerza con la cual dos variables estan correlacionadas. Al ser un coeficiente _r_ no tiene unidades y sus valores oscilan entre $-1 =< r =< 1$. Valores cercano a 1 o -1 indican una fuerte relacion entre ambas variables y valores cercanas a 0 una relacion debil.

Las suposiciones para correr una correlacion de pearson son el nivel de las medidas que los pares esten relacionados, idealmente que no hayan outliers, y que la correlacion se vea linear

El nivel de las medidas en el caso de pearson se refiere a que ambas variables deben de ser continuas. Si una o ambas variables son ordinales entonces se debe de llevar acavo otro tipo de correlacion como Spearman.


```{r}

# Revisar suposiciones
plot(FlujosC$CH4.conc,FlujosC$CO2.conc)

# Sacar correlacion
cor.test(FlujosC$CH4.conc,FlujosC$CO2.conc)
cor.test(FlujosC$CH4.conc,FlujosC$CO2.conc,
         method = "pearson") #tambien se puede revisar asi, en el metodo podemos poner si queremos pearson o spearman.

####ALGO IMPORTANTE ES QUE SI NO SE PARA QUE ES UN CODIGO EN LA CONSOLA LE PONGO SIGNO DE INTERROGACION Y EL CODIGO Y EN LA CONSOLA DERECHA INFERIOR ME SALE PARA QUE SIRVE 
```


Es importante notar que en este caso tambien hay un p-value y para que esta correlacion sea significativa p<0.05. Incluso aunque _r_ sea bajo si p<0.05 esta relacion es significativa

# Regresion

Una vez que sabemos si dos variables cuantitativas estan correlacionadas el siguiente paso es expresar esta relacion en la forma de una ecuacion. Usando esta ecuacion uno podria calcular una de las variables si se sabe la otra 

```{r}
# lm: modelo lineal y~x
summary(lm(FlujosC$CO2.conc~FlujosC$Minute)) # aqui estoy midiendo una regresion lineal de la liberacion de co2 por minuto 

plot(FlujosC$Minute,FlujosC$CO2.conc) #x mi variable y y lo que codifico
abline(lm(FlujosC$CO2.conc~FlujosC$Minute)) # aqui te dice que le agregue una linea lineal en el grafico 

```

En este caso el modelo linear da los valores de la interseccion en y la pendiente para nosotros escribir la ecuacion: y= mx + b

  __CO2= -2.10941x+215.72__

En la ecuacion la pendiente nos senala que cada minuto el CO2 disminuye _-2.1 ppm_

Asimismo, esta funcion nos da el valor de __r2= 0.97__ (no confundir con r) que en este caso significa que el 97% de la variacion en CO2 es explicada por el tiempo. Esta ecuacion tambien cuenta con su nivel de significancia y en este caso es menor a p<0.0001


```{r}
summary(lm(FlujosC$CH4.conc~FlujosC$Minute))
# CH4= 0.41 minute + 12.29 osea que es la formula y= mx + b
```

##porque en este caso no corrimos este: 

```{r}
summary(lm(FlujosC$CH4.conc~FlujosC$CO2.conc))
```
si bien el codigo corre de igual menera en este caso ambas son varialbes dependientes por lo que fueron variablews que medimos. Al haberlas medido no seria tan util predecir una usando la otra ya que tendriamos que de iugal manera medir una de ellas. 

# Ejercicio 

Busquen o creen un set de datos en el que se pueda analizar una correlacion y regresion lineal. Justifiquen su respuesta 

```{r}
datos <- data.frame(vueltas= c(3,6,9,12,15), tiempo = c(14,19,23,27,26))

cor.test(datos$tiempo,datos$vueltas)

summary(lm(datos$tiempo~datos$vueltas))

#  y = mx + b
# Tiempo = 1.06 (vueltas) + 12.2

```
