==> NoteBook/08_Analisis no parametricos.Rmd <==
---
title: "Analisis no parametricos"
output: html_notebook
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Ctrl+Shift+Enter*. 

```{r}
plot(cars)
```

Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Ctrl+Alt+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Ctrl+Shift+K* to preview the HTML file).

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.



# Introduccion

Como vimos antes nuestros datos no van a poder ser analizados siempre por analisis parametricos ya sea porque no se pueden normalizar o porque la naturaleza de los datos no lo permite. Para ello vamos a aprender varios tests no parametricos. Recuerden que siempre es mejor tratar de hacer todo lo posible para usar test parametricos antes de recurrir a estos test ya que son mas restrictivos y no tan sensibles como los tests parametricos. 

```{r}
# Librerias

library(dplyr)
library(tidyverse)
library(car)
library(rstatix)
library(PMCMRplus)

```

```{r}
Estanques <- read.csv("~/RStudio/CursoInnovak/Materiales/BRW.ponds.csv")

Exudados <- read.csv("~/RStudio/CursoInnovak/Materiales/Exudados.csv")

Comp_suelos <- read.csv("~/RStudio/CursoInnovak/Materiales/Soil_Comparison.csv")

Set_datos <- read.csv("~/RStudio/CursoInnovak/Proyecto_1_RMarkdown/Set de datos proyecto.csv")
```

# Kruskal Wallis 

este metodo se usa como sustituto a la _ANOVA de una via_ ya que no necesita normalidad de distribucion en la poblacion ni homogeneidad de varianza 

```{r}
# Primero revisar normalidad 
shapiro.test(Estanques$DOC)
hist(Estanques$DOC) # si son normales 

# Homogeneidad de varianza con el levene test 
leveneTest(DOC ~ Region, data = Estanques)
# la cumple? NO

# Kruskal Wallis 
kruskal.test(DOC ~ Region, data = Estanques)

# Comparar con ANOVA
est_anova <- aov(DOC ~ Region, data = Estanques)
Anova(est_anova)

```
Si lo comparamos con el ANOVA el Kruskal Wallis es mas estricto y por lo tanto a veces puede afectar la sgnificancia estadistica. 

## Wilcoxon Rank

Como con la ANOVA Kruskal Wallis nos dice4 si hay o no diferencias pero no CUALES son esas diferencais. Para ello usamos el equivalente no parametrico al test de _Tukey HSD_. El pairwise Wilcoxon test es un post hoc test que se puede usar como sustitucion de Tukey. 

```{r}

pairwise.wilcox.test(Estanques$DOC,
                     Estanques$Region,
                     p.adjust.method = "none")
# Comparar con Tukey
TukeyHSD(est_anova)
```
## Que es el p.adjust method?

Ajusta los p-value cuando existen multiples comparaciones usando diferentes metodos. Existen varios metodos .......link....... y aunque su explicacion matematica es algo complicada es importante que conozcan su aplicacion. La funcion principal de estos metodos es ajustar p-value tomando en cuenta las comparaciones para evitar que se inflen demasiados y evitar falsos positivos. El metodo mas comun y tambien mas estricto es _bonferroni_ otro metodo menos estricto es _fdr_ o false discovery rate. Para aprender mas sobre que metodo usar pueden checar esta pagina: 
........pagina.....

```{r}
pairwise.wilcox.test(Estanques$DOC,
                     Estanques$Region,
                     p.adjust.method = "bonferroni")

pairwise.wilcox.test(Estanques$DOC,
                     Estanques$Region,
                     p.adjust.method = "fdr")

# compara con Tukey
TukeyHSD(est_anova)
```
## Dunn test 

```{r}
dunn.test::dunn.test(Estanques$DOC,
                     Estanques$Region,
                     method = "bonferroni")
```
### Ejercicio 

En el set de datos EXUDADOS hay 3 compuestos con variables dependientes. Determinen cual de ellos no se puede analizar meidante test parametricos y analizarlo usando Kruskal Wallis para ver si existen diferencias significativas entre trateientos. De ser asi raelizar un post hoc test para ver cuales son las diferencias. 

```{r}
shapiro.test(Exudados$Valarate_C18P) 
hist(Exudados$Valarate_C18P)
qqnorm(Exudados$Valarate_C18P)

shapiro.test(Exudados$Chorismate_C18N)
hist(Exudados$Chorismate_C18N)
qqnorm(Exudados$Chorismate_C18N)

shapiro.test(sqrt(Exudados$Chorismate_C18N))
hist(sqrt(Exudados$Chorismate_C18N))
qqnorm(sqrt(Exudados$Chorismate_C18N))

shapiro.test(Exudados$Glyoxalate_HILN)
hist(Exudados$Glyoxalate_HILN)
qqnorm(Exudados$Glyoxalate_HILN)

shapiro.test(log(Exudados$Glyoxalate_HILN))
hist(log(Exudados$Glyoxalate_HILN))
qqnorm(log(Exudados$Glyoxalate_HILN))

# Homogeneidad de varianza con el levene test 
leveneTest(Valarate_C18P ~ Treatment, data = Exudados)
# la cumple? si

leveneTest(Chorismate_C18N ~ Treatment, data = Exudados)
# la cumple? si

leveneTest(Glyoxalate_HILN ~ Treatment, data = Exudados)
# la cumple? no

# Kruskal Wallis 
kruskal.test(Glyoxalate_HILN ~ Treatment, data = Exudados)


# Comparar con ANOVA
exu_anova <- aov(Glyoxalate_HILN ~ Treatment, data = Exudados)
Anova(exu_anova) #hay diferencia significativa 


# Wilcoxon
pairwise.wilcox.test(Exudados$Glyoxalate_HILN,
                     Exudados$Treatment,
                     p.adjust.method = "none")

pairwise.wilcox.test(Exudados$Glyoxalate_HILN,
                     Exudados$Treatment,
                     p.adjust.method = "fdr")

# Comparar con Tukey
TukeyHSD(exu_anova)

```


# Friedman test 

Como pueden ver en el caso de los exudados el test Kruskal Wallis no nos permite comparar interacciones entre dos variables independientes como con una ANOVA de dos vias. En este caso el equivalente no parametrico es el test Friedman. Sin embargo, este test es sumamente restrictivo en los set de datos que pueden ser analizados y a la vez es major usar kriskal wallis por separado en cada variable independiente. 

```{r}
data("selfesteem", package = "datarium")
head(selfesteem,3)
# los datos tienen que estar balanceados 

selfesteem <- selfesteem %>%
  gather(key = "time", value = "score", t1, t2, t3) %>%
  convert_as_factor(id,time)
head(selfesteem, 3)

friedman.test(score ~ time |id, data = selfesteem)

# Post hoc test equivalente a Tukey
frdAllPairsNemenyiTest(score ~ time |id, data = selfesteem)
```
### Ejercicio final 

La mayor parte de ustedes tuvo alguna variable en sus proyectos que no se podria analizar con test parametricos. Usando lo que aprendimos hoy elijan una de esas variables y analicenla usando Kruskal Wallis con cada variable independiente por separado. 


```{r}
shapiro.test(Set_datos$Long_Raiz) 
hist(Set_datos$Long_Raiz)
qqnorm(Set_datos$Long_Raiz)

shapiro.test(Set_datos$Peso_sem_gr)
hist(Set_datos$Peso_sem_gr)
qqnorm(Set_datos$Peso_sem_gr)

shapiro.test(log(Set_datos$Long_Raiz))
hist(log(Set_datos$Long_Raiz))
qqnorm(log(Set_datos$Long_Raiz))

shapiro.test(sqrt(Set_datos$Long_Raiz))
hist(sqrt(Set_datos$Long_Raiz))
qqnorm(sqrt(Set_datos$Long_Raiz))

# Kruskal Wallis 
kruskal.test(Long_Raiz ~ Tratamiento, data = Set_datos)
kruskal.test(Long_Raiz ~ Dia, data = Set_datos)


# Wilcoxon


pairwise.wilcox.test(Set_datos$Long_Raiz,
                     Set_datos$Dia,
                     p.adjust.method = "none")

pairwise.wilcox.test(Set_datos$Long_Raiz,
                     Set_datos$Dia,
                     p.adjust.method = "fdr")

pairwise.wilcox.test(Set_datos$Long_Raiz,
                     Set_datos$Dia,
                     p.adjust.method = "bonferroni")

```
