==> NoteBook/09_Visualizacion de datos usando ggplot2.Rmd <==
---
title: "Visualizacion de datos usando ggplot2"
output: html_notebook
---


# Introduccion

Graficar con R tiene varias posibilidades. Hasta este momento todos los graficos que hemos hecho han sido con las funciones base. Si bien con las funciones base tambien se pueden hacer graficos mediamente decentes, la base para graficar en RStudio ahora es el paquete ggplot2.

El nombre de _ggplot2_ viene de la _Gramatica de Graficar_ y creo que es importante de mencionar ya que la sintaxis que usa es muy parecida a bloques de gramatica y aunque no lo crean mucho mas facil de entender que R base. 

1.- Datos: los datos en bruto que quieres graficar.

2.- Geometrias geom_: las figuras que van a representar los datos 

3.- Estetica aes(): que variables voy a graficar y que tan bonito se va ver nuestro grafico y sus objetos estadisticos. Aqui es especifica posicion, color, tamaño, forma y transparencia. 

4.- Escala scaole_: las dimensiones de tu grafico y donde van a ser "dibujados" tus datoa. 

5.- Transformaciones estadisticas stat_: resumentes estadisticos que incluir en el grafico como quantiles, curvas y sumas 

6.- Coordenadas coor_

7.- Facetas facet_: el arreglo de los datos en una red de graficos 

8.- Te,as theme(): va a ser su mejor amigo! En general cambia de una manera rapida varios elementos visuales del grafico (fondo no datos)

```{r}
# Librerias 

library(ggplot2)
library(tidyverse)

# Data 
data("airquality")
View(airquality)

FlujosC <- read.csv("~/RStudio/CursoInnovak/Materiales/flujos.csv")

```

# Sintaxis basica de un ggplot 

La sintaxis para ggplot2 nos entienda es un poco diferente de R base. 
De los elementos que les explique arriba un grafico base necesita que le especifiquen 3 cosas: 

* Datos: lo primero que definimos al abrir el comando ggplot()
* Estetica: dentro de ese mismo compando entra aes() aqui es donde vamos a decirles al grafico que es lo que vamos a graficar es decir le vamos a especificar las variables que queremos que nos grafrique. 
* Geometria 

```{r}
g <- ggplot(airquality, aes(x= Day, y= Temp)) #si le ponemos el objeto no va a aparecer tendriamos que poner el la palabra o letra en la consola de abajo la letra y nos aparecera el grafino en la consola derecha inferior 
```
Porque nos sale vacio el grarfico?
La ventaja de ggplot es que se puede guardar como un objeto al cual se le pueden ir agregando cosas: 

```{r}
# agregados geometricos 

g + geom_point()
g + geom_line()
g + geom_line() + geom_point()
```
Como pueden ver con los mismos datos se puede hacer un grafico totalmente diferente. La clave esta en saber que bloques usar para constuirlo! 

## Modificando las propiedades de una geometria 

```{r}
g + 
  geom_point(color= "pink", shape= "diamond", size= 2)+
  geom_line(color= "darkblue", linetype= "dotted", lwd= .3)
```
A partir de esta parte quiero que exploren, cambien coloresk, modifiquen formas. Como con cualquier grafico no hay forma correcta o incorrecta de hacerlo. Tampoco es necesrio que se memoricen TODOS los colores, formas o tipos de fila: 

## Temas

```{r}
theme_set(theme_bw())

g + geom_point(color= "blue")

g + geom_point(color= "azure")+ theme_dark()
```

como pueden ver hay mas temas que ideas y dentro de un tema hay muchos elementos que se pueden modificar. El limite es su tiempo, ganas y la necesidad. 

## Ejercicio 

Usando los datos de ejemplo, creen que temperatura y radiacion solar esten correlacionados de alguna manera? grafiquen con lo que hemos aprendido y diganme su opinion.  

```{r}
# Revisar suposiciones
plot(airquality$Solar.R,airquality$Temp)

# Sacar correlacion
cor.test(airquality$Solar.R,airquality$Temp)
cor.test(airquality$Solar.R,airquality$Temp,
         method = "pearson") # si esta correlacionado...porque no se pero si esta. 


# grafico 
grafico <- ggplot(airquality, aes(x= Solar.R , y= Temp))

# agregar geometricos 

grafico + geom_line() + geom_point()

# modificaciones 
grafico + 
  geom_point(color= "gray", shape= 2, size= 2)+
  geom_line(color= "darkorange", linetype= "dotted", lwd= .5)+
  theme_minimal()

```

PARA AGR4EGAR UNA LINEA DE CORRELACION SERIA LA OPCION DE
geom_abline()

## Cambiando los ejes 

```{r}
ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "maroon4") +
  labs(x= "Day", y= "Temperature (°F)")


# agregando superscripts 
ggplot(airquality, aes(x= Day, y= Temp))+
  geom_point(color= "darkblue") +
  labs(x= "Day",
       y= expression(paste("Temperature (°F)"
                           ^"(Porfa  cambien a Celsius")))

# agregando subscrit 
ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(size= 2, fill= "cyan3", 
               color= "chartreuse4", shape= 21) +
  xlab("Day")+
  ylab(bquote("Temperature (°F)"["Porfa cambien a Celsius"]))

# Cambiando tamaños 

ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "darkorchid1") +
  labs(x= "Day", t= "Temperature (°F)") +
  theme(axis.title.x = element_text(size = 15),
        axis.title.y = element_text(size= 15))


#vs. en todo el grafrico 

ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "darkorchid1") +
  labs(x= "Day", t= "Temperature (°F)") +
  theme_light(base_size = 15) # esta es la parte que lo modifica 

```

Y de ahi de nuevo el cielo es el limite. Ademas de las modificaciones a los titulos de los ejes tambien se pueden modificar la estetica de la escala como tal: 

```{r}
ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "firebrick1") +
  labs(x= "Day", y= "Temperature (°F)") +
  theme(axis.text = element_text(color= "maroon",
                                 size= 12,
                                 face= "italic"))

# o el angulo del texto 

ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "firebrick1") +
  labs(x= "Day", y= "Temperature (°F)") + #encabezados de la grafica 
  theme(axis.text = element_text(color= "maroon",
                                 size= 12,
                                 face= "italic"),
        axis.title.x = element_text(angle = 1,
                                    vjust = 1, #ajuste vertical
                                    hjust = 1, #ajuste horizontal 
                                    size = 12))


# o de donde a donde va tu escala 

ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "tan2") +
  labs(x= "Day", y= "Temperature (°F)") + #encabezados de la grafica 
  ylim(c(0,100))
 
```

# Modificando los titulos 

```{r}
ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "firebrick1") +
  labs(x= "Day", y= "Temperature (°F)") +
  ggtitle("Calidad del aire") #para agregar titulos  

# ocupas mas que un solo titulo?

ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "dodgerblue") +
  labs(x= "Day", y= "Temperature (°F)",
       title = "Calidad del Aire",
       subtitle = "odio los grados Fahrenheit",
       caption = "Datos de ejemplo", #como si fuera un pie de pagina 
       tag = "Fig.1")

# Para mover el titulo de lugar osea mas centrado...

ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "firebrick1") +
  labs(x= "Day", y= "Temperature (°F)") +
  ggtitle("Calidad del aire")+
  theme(plot.title = element_text(hjust = 0.5, size= 25)) #ubicacion del texto, el uso de plot.title es para mover de lugar los titlos ya sea de las lineas, margenes, todo lo relacionado a la grafica 

# modificando la estatica del titulo

ggplot(airquality, aes(x= Day, y= Temp)) +
  geom_point(color= "firebrick1") +
  labs(x= "Day", y= "Temperature (°F)") +
  ggtitle("Calidad del aire")+
  theme(plot.title = element_text(face = "bold",size = 14)) #face es para cambiar tipo de texto 

                            
```


# Ejercicio final 

Ahora si haran su primer grafico desde cero, usando los datos de FlujosC quiero que me grafiquen la concentracion de CH4 por minuto. El grarfico debe de incluir cada uno de los siguientes elementos: 

1.- Por lo menos un cambio dentro de la geometria del grafico 
2.- Etiquetas de los ejes (y quiero que el 4 este en subscript)
3.- Modificacion de la escala de los ejes 
4.- Un titulo con por lo menos una modificacion en formato 

```{r}
graf.FlujoC <- ggplot(FlujosC, aes(x= CH4.conc , y= Minute))


ggplot(FlujosC, aes(x= Minute, y= CH4.conc))+
  geom_point(color= "#00cd66") +
  labs(x= "Tiempo (min)",
       y= bquote("Concentracion de CH"[4]),
       title = "FlujosC",
       caption = "Datos compartidos por Isabel")+
  theme(axis.title.x = element_text(size = 20),
        axis.title.y = element_text(size = 20),
        plot.title = element_text(hjust = 0.5, size = 25))
         



```
