==> NoteBook/13_Phyloseq.Rmd <==
---
title: "Phyloseq: analizar microbiomas con R"
output: html_notebook
---


Nuestro pipeline de DADA2 nos da como resultado 2 tablas: el conteo de las secuencias (abundancias) y la tabla de asignacion taxonomica. Ademas de ello, tenemos los datos de las caracteristicas de nuestras muestras, es decir los metadatos: Tipos de muestra, lugar, tratamientos, cultivos, etc. Tratar de balancear toda esta informacion al momento de analizar nuestros datos puede ser una tarea dificil por lo que R tiene un paquete para ello: __Phyloseq__

Phyloseq es un paquete que ayuda a importar, almacenar, analizar y presentar graficamente datos de secuenciacion filogenetica que ya ha sido agrupada en OTUs o ASVs. Especialmente ayuda cuando ya esta asociada a datos de la muestra, arboles filogeneticos,y tablas de asignacion taxonomica agrupandolos en un solo objeto. Usando este objeto se pueden implementar otras herramientas disponibles en R para el analisis filogenetico.

Algunos ejemplos:

* Importar datos de denoising pipelines como DADA2, pero existen mas (es lo que haremos hoy)

* Analisis de diversidad: metodos de distancias y ordinacion

* Graficos de microbioma usando ggplot 2 (osea despues se pueden modificar con lo que ya aprendimos)

* Manejos, calculos y modificaciones de objetos phyloseq


# Instalacion

## 1. Bioconductor 

```{r}
source('http://bioconductor.org/biocLite.R') # este no sirve 
biocLite('phyloseq') # esto es para versiones mas viejas de R 
packageVersion('phyloseq')

BiocManager::install("phyloseq")

library(phyloseq)

load("taxa_ch.RData") # Tabla de identificacion taxonomia 
load("seq_conteos.RData") # Tabla de abundancia ASVs

# Metadados: nosotros haremos la tabla 

metadatos <- data.frame(Tratamientos= c(rep("Tratamiento", 3)),
                        Tipo_de_suelo= c(rep("Suelo", 3)))
row.names(metadatos) <- sample_names

```


## Preparando los datos 

Si bien nuestras matrices ya estan listas para usarse hay algunas modificaciones que son necesarias para crear nuestro objeto de phyloseq

### Abundancia de ASVs

```{r}
sample_names <- c("CH1", "CH2", "CH3")
row.names(seqtab.nochim) <- sample_names

```

Sin embargo, nuestras columnas se siguen llamando como nuestras secuencias. Esto quiere decir que tiene nombres super largos y que al momento de analizar se nos va a complicar visualizar

```{r}
# Generar codigos aleatorios como nombres de las columnas 
rando <- function(n = 5000) {
  a <- do.call(paste0, replicate(5, sample(LETTERS, n, TRUE), FALSE))
  paste0(a, sprintf("%04d", sample(9999, n, TRUE)), sample(LETTERS, n, TRUE))
}

dim(seqtab.nochim) # para revisar cuantas filas y columnas tengo en mi tabla 
seqnames <- rando(1609) # este numero lo sacamos del seqname entre []

# Guardar las secuencias en otro objeto antes de borrarlas 
Secuencias <- colnames(seqtab.nochim)
write.csv(Secuencias, "~/RStudio/CursoInnovak/Materiales/Secuencias.csv")


# Cambiando nombre con codigo
colnames(seqtab.nochim) <- seqnames

### O nombrarlas de manera ordinal (seria otra opcion)
colnames(seqtab.nochim) <- c(paste0("Seq_", as.character(1:1609)))

# Guardar la tabla (dependiendo como lo quiera guardar con codigo o con numero osea de manera ordinal)
write.csv(seqtab.nochim,
          "~/RStudio/CursoInnovak/Materiales/Tabla_ASVs.csv")
```

### Tabla de asingancion taxonomica 

Para que nuestras tablas se puedan "unir" y analizarse juntas dentro del objeto phyloseq necesitamos tener los mismos nombres para nuestas variables. Por ello ahora tenemos que cambiar los nombres de las filas de nuestra tabla de taxa (que son la secuencias que acabamos de quitar) al codigo o nombre que le dimos en la parte de arriba: 

```{r}
row.names(taxa) <- seqnames #esta es para si nos quedamos con los nombres de codigos 

row.names(taxa) <- c(paste0("Seq_",as.character(1:1609))) # esta es para si te quedas con la opcion de nombrarlos por numero 

# Cuando decides como lo llamaras lo Guardas...
write.csv(taxa, "~/RStudio/CursoInnovak/Materiales/taxa.csv")

```


Ahora si, todo listo para armar nuestro objeto y usarlo para todo tipo de analisis 

```{r}
suelo_ch <- phyloseq(otu_table(seqtab.nochim, taxa_are_rows = FALSE), #aqui le ponemos que no porque nuestras no estan en las filas estan en las columnas
                     sample_data(metadatos),
                     tax_table(taxa))

# y o se olviden de guardar 
save(suelo_ch, file = "Suelos_chiapas.RData")
```
