==> NoteBook/14_DADA2 - ITS.Rmd <==
---
title: "DADA2 para ITS"
output: html_notebook
---


# Introduccion 

Otra variacion importante de DADA2 es como se puede aplicar al analisis de ITS. De la misma manera que la region 16S se usa para identificar bacterias, la region ITS se usa para identificar hongos. La diferencia mas grande entre ambas regiones es que el tamano de la region ITS es altamente variable por lo que el pipeline debe modificarse un poco.

En este caso el punto de partida son archivos fastq pareados que fueron secuenciados por Illumina qe fueron separados (demultiplexed) por muestra y sin barcodes. Al igual que con 16S el resultado es una tabla de ASVs y otra asignando taxonomia usando la base de datos UNITE. 

__La adicion clave al pipeline comparado con 16S es la identificacion, verificacion de la orientacion y remocion de los primers__

## Porque es necesaria esta modficacion

A diferencia del gen 16S rRNA, la región ITS tiene una longitud muy variable. Las regiones ITS1 e ITS2 comúnmente amplificadas varían entre 200 y 600 bases de longitud. Esta variación de longitud es biológica, no técnica, y surge de las altas tasas de inserciones y eliminaciones en la evolución de esta región genética menos conservada.

La variación de longitud de la región ITS tiene consecuencias significativas para los pasos de filtrado y recorte del flujo de trabajo estándar DADA2. En primer lugar, el truncamiento a una longitud fija ya no es apropiado, ya que ese enfoque elimina las variantes ITS reales con longitudes más cortas que la longitud del truncamiento. En segundo lugar, la eliminación del primer se complica por la posibilidad de que algunas, pero no todas, las lecturas se extiendan al primer opuesto cuando la región ITS amplificada es más corta que la longitud de la lectura.

En el flujo de trabajo estándar 16S, generalmente es posible eliminar 
los primers en el proceso de filtrado: 

(filterAndTrim(..., trimLeft=(FWD_PRIMER_LEN, REV_PRIMER_LEN))) 

ya que solo aparecen al comienzo de las lecturas. y tienen una longitud fija. Sin embargo, la complejidad de la región ITS requiere el uso de herramientas externas. Aquí vamos a usar _cutadapt_ para la eliminación de primers.

# Librerias

```{r}
library(dada2)

# Instalar 
BiocManager::install("ShortRead")
BiocManager::install("Biostrings")

library(ShortRead)
packageVersion("ShortRead")
library(Biostrings)
packageVersion("Biostrings")
```
# Seleccion y preparacion de datos 

Como en este caso ustedes aun no tienen muestras de ITSs vamos a usar los datos de mi tesis: 

```{r}
### Camino ###

path <- "~/RStudio/CursoInnovak/Seq ITS/"
list.files(path) # este es para ver la lista/archivos de lo que tiene la carpeta, esto es para saber si estamos en la carpeta deseada 


## Forward y reverse reads ##

fnFs <- sort(list.files(path, pattern = "_1.fastq.gz", full.names = TRUE))

fnRs <- sort(list.files(path, pattern = "_2.fastq.gz", full.names = TRUE))

```

# Identificar los primers 

Al momento de entregarme mis secuencias me dieron un archivo con barcodes, primers y linkers y de ahi lo saque: 

```{r}

FWD <- "GGAAGTAAAAGTCGTAACAAGG"
REV <- "GCTGCGTTCTTCATCGATGC"

```


Ahora tenemos que checar la orientacion de todos los primers: 

```{r}
### Funcion predeterminada ###

allOrients <- function(primer) {
  # Crear todas las orientaciones de esas secuencias
  require(Biostrings)
  dna <- DNAString(primer)  # Para usar Biostrings usa DNAString objects no vectores
  orients <- c(Forward = dna,
               Complement = Biostrings::complement(dna),
               Reverse = Biostrings::reverse(dna),
               RevComp = Biostrings::reverseComplement(dna))
  return(sapply(orients, toString))  # De regreso a vector
}

FWD.orients <- allOrients(FWD)
REV.orients <- allOrients(REV)
FWD.orients #te la las cuatro posibilidades de varacion de la secuencia 

```
