==> NoteBook/15_Herr_Phyloseq.Rmd <==
---
title: "Herramientas de phyloseq"
output: html_notebook
---

```{r}
library(phyloseq)

```


Ejercicio de recapitulacion; como guardar en un phyloseq

```{r}
# paso 1: subir tabla de dada
load("taxa_ch_proy2_1_esp.RData") #identificar taxonomia 
load("seq_conteos_proy2.RData") # abundancia ASVs

# paso 2: metadados 
metadatos <- data.frame(Tratamiento= "Bioestimulante",
                        Suelo= "No salino")
row.names(metadatos) <- "S81"

# paso 3: renombar las muestras 
row.names(seqtab.nochim_proy2) <- "S81"

# paso 4: cambiar nombres de secuencias y crar secuencia en objeto
rando <- function(n = 5000) {
  a <- do.call(paste0, replicate(5, sample(LETTERS, n, TRUE), FALSE))
  paste0(a, sprintf("%04d", sample(9999, n, TRUE)), sample(LETTERS, n, TRUE))
} # esta es una funcion y no se modifica 

dim(seqtab.nochim_proy2) # para revisar cuantas filas y columnas tengo en mi tabla, en parentesis va como se llama nuestra tabla 
seqnames <- rando(3961) # el resultado de la funcion anterior va en este parentesis, se encuentra en corchetes en el ambiente 

# Guardar las secuencias en otro objeto antes de borrarlas 
Secuencias <- colnames(seqtab.nochim_proy2)
write.csv(Secuencias, "~/RStudio/CursoInnovak/Materiales/Secuencias_ejercicio_phyloseq.csv")

# cambiar nombre con codigo
colnames(seqtab.nochim_proy2) <- seqnames

# paso 5: Cambiando nombre a las secuencias en taxa
row.names(taxa_proy2_esp_2) <- seqnames

# paso FINAL: 
vid_bio <- phyloseq(otu_table(seqtab.nochim_proy2,
                              taxa_are_rows = FALSE),
                    sample_data(metadatos),
                    tax_table(taxa_proy2_esp_2))

save(vid_bio, file = "VidS81.RData")

```

Introducción 

Como vimos antes de irnos phyloseq nos ayuda a integrar todos nuestros datos en un objeto para poder analizarlos. La clase vamos a ver como phyloseq nos deja modificar estos objetos para visualizarlos mejor y seguir con los diferentes tipos de analisis usando mis datos de maestria 

```{r}
load("VidS81.RData") #siempre empezar con un objeto de phyloseq, en este caso ya no corro este codigo proque ya lo tengo en el ambiente (mi archivo se llama VidS81.RData pero en el ambiente le escribimos vid_bio)
```

# las primeras visualizaciones de nuestros datos 

```{r}
# REDES; como estan interactuando nuestras muestras
net <- make_network(vid_bio, "samples", max.dist = 0.3)
plot_network(net,vid_bio, color = "Treatment", shape = "Plant_Type",
             line_weight = 0.3, label = NULL)

plot_bar(vid_bio, fill = "Phylum")
plot_heatmap(vid_bio, taxa.label = "Phylum") # no corre porque es solo una muestra
```
Si bien con estos datos ya podemos realizar graficos que nos acercan mas al analisis el hecho es de que si siempre es conveniente pre-procesar los datos antes de cualquier grafico. Para ello phyloseq nos ofrece varias herramientas: 

# Pre-procesamiento de datos

## Filtrado

```{r}
# Por que tres muestras? 
PS_filtered <- filter_taxa(vid_bio, #objeto de phyloseq
                           function(OUT) sum(OUT) > 2, #condicion o funcion 
                           TRUE) # el true es para saber si quieres cortar o no y crear un nuevo objeto, siempre dejarlo en TRUE
#este es un proceso de filtrado por numero de muestras

#En este caso son 2 muestras porque fueron 2 replicas, para que sea significativo y no por obra de Dios

# Remover taxa on identificada
PS_filtered <- subset_taxa(PS_filtered, # objeto
                           !is.na(Phylum)) #condicion, de los que no tengan nada en phylum me los vas a quitar; es lo que le estamos diciendo en este codigo

#muestras vid_bio = 3961
#primer filtrado = 2916
#segundo filtrado = 2822
```

### Prune vs Subset

```{r}
# Subset, usa los datos de la tabla de taxonomia (datos de identificacion), quito phylum porque filtra tax_table
Actino <- subset_taxa(vid_bio, #objeto
                      Phylum=="Actinobacteriota") #la condicion puede ser cualquier categoria/nivel de la jerarquia taxonomica por ejemplo genero,etc
#muestra actino = 873

# Prune, quita muestras, filtra el out_table
Actino <- prune_samples(sample_sums(Actino)>=50, #condicion
                        Actino) #objeto

```

## Union o merge

```{r}
# Uniendo muestras 
Tratamientos <- merge_samples(vid_bio, #objeto
                              "Tratamiento") #condicion de metadatos
#en este caso no se puede porque solo tenemos una muestra 

# Uniendo taxa; no es tan util
Actino_m <- merge_taxa(Actino, taxa_names(Actino)[1:5]) #por numero

# Uniendo por jerarquia; es la mas util (glom de aglomerar)
PS_glom <- tax_glom(PS_filtered, #objeto (podemos usar el original o el filtrado)
                    taxrank = "Genus", #nivel jerarquia
                    NArm = FALSE)

# Uniendo phyloseqs ; unir varios objetos de phyloseq
merge_phyloseq(Actino, Desulf) #entre parentesis son los objetos que vamos a unir 
```


### Ejercicio

Unir las muestras de todos 
```{r}
# Como todos se llamaban igual, al momento de subir tuvimos que renombrarlas antes de subir la siguiente
load("VidS81.RData")
V1 <- vid_bio

load("VS82.RData")
V2 <- vid_bio

load("VidS83.RData")
V3 <- vid_bio

load("Vid_S85.RData")
V5 <- vid_bio

vid_bio <- merge_phyloseq(V1, V2, V3, V5) # al usar marge fijarse que los metadatos coincidan 

save(vid_bio, file = "Vid_ejemplos.RData")
load("Vid_ejemplos.RData")

# Modificando la tabla de metadatos 
vid_bio@sam_data$Suelo[is.na(vid_bio@sam_data$Suelo)] <- "Salino"
vid_bio@sam_data$Tratamiento[vid_bio@sam_data$Tratamiento == "Sin bioestimulante"] <- "Control" # esto es para modificar si tuvimos un error de dedo al escribir algo, en este caso le estamos diciendod que en el objeto vid_bio en la seccion del tratamiento cambiamte (por eso son ==) el sin bioestumulante a control 
vid_bio@sam_data <- vid_bio@sam_data[,1:2] #Recordar que cosas entre corchetes [] la coma nos dice si son filas o columnas "[filas,columnas"]

save(vid_bio, file = "Vid_ejemplos.RData") #aqui ya esta guardado con las modificaciones de los errores que tuvimos 
```

```{r}

net <- make_network(vid_bio, "samples", max.dist = 2) #modificar la distancia cuando no hayy conexiones, en este caso se cambio el max.dist porque las muestras estan muy separadas 
plot_network(net,vid_bio, color = "Suelo", shape = "Tratamiento",
             line_weight = 0.3, label = NULL)


```

## Abundancia relativa

DADA2 nos da abundancias absoluta, osea si conto 10 mil secuencias pues son 10 mil secuencias. Las cuales pasamos a nuestro objeto de phyloseq.
Por lo tanto, la abundancia absoluta es un conteo TOTAL de la especies e individuos dentro de cada especie (las relativas se usan para graficos y cuando queremos comparar graficas), 

La abundancia relativa nos ayuda a comparar diferentes muestras, osea normaliza nuestras muestras para poder compararlas entre si. Para ello, usa proporciones siendo 1 el 100% de nuestro conteo por muestras. De esta manera podemos comparar la composicion de nuestras muestras aunque no tengamos valores totales o absolutos iguales. (no importa que de una muestra tenga mas que otra muestra).

Para transformar de abundancia ABSOLUTA a RELATIVA o cualquier otro tipo de calculo para nuestras muestras se usa la siguiente funcion de phyloseq : 

```{r}
Psoil_rel <- transform_sample_counts(PS_filtered, #objeto a tranformar
                                     function(x) x / sum(x)) # x es mi conteo
```

# Reexplorarmos graficos 

```{r}
# 1. Grafico de barras
Top_phyla <- tax_glom(Psoil_rel,taxrank = "Phylum", NArm = FALSE)
Top_phyla <- prune_taxa(names(sort(taxa_sums(Top_phyla), TRUE)[1:10]),
                        Top_phyla) # el objeto que voy a modificar 

plot_bar(Top_phyla, fill = "Phylum")
plot_heatmap(Top_phyla, taxa.label = "Phylum")
```
